spp.deployment.kubernetes/requirements.txt
pyyaml >= 6.0.1, < 7.0
